在发布事件时获取traceId
