+++ src/main/java/com/tehang/common/utility/event/DomainEvent.java
@@ -1,12 +1,9 @@
 package com.tehang.common.utility.event;
 
-import brave.Tracer;
-import com.tehang.common.utility.ApplicationContextProvider;
 import com.tehang.common.utility.JsonUtils;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.NoSuchBeanDefinitionException;
 
 import java.util.UUID;
 
@@ -57,39 +54,6 @@ public abstract class DomainEvent {
   public DomainEvent(String eventType) {
     this.key = UUID.randomUUID().toString();
     this.eventType = eventType;
-    this.traceId = getCurrentTraceId();
-  }
-
-  /**
-   * 获取当前上下文中的traceId
-   */
-  private static String getCurrentTraceId() {
-    // 获取当前的springContext
-    var applicationContext = ApplicationContextProvider.getApplicationContext();
-    if (applicationContext == null) {
-      return null;
-    }
-
-    Tracer tracer;
-    try {
-      // 获取当前上下文的tracer
-      tracer = applicationContext.getBean(Tracer.class);
-    }
-    catch (NoSuchBeanDefinitionException ex) {
-      // 未启用Tracer时，直接返回null
-      log.debug("cannot found Tracer bean");
-      return null;
-    }
-
-    if (tracer.currentSpan() != null && tracer.currentSpan().context() != null) {
-      var traceContext = tracer.currentSpan().context();
-      var traceId = traceContext.traceId();
-      if (traceId != 0) {
-        log.debug("Exit getCurrentTraceId: {}", traceId);
-        return String.valueOf(traceId);
-      }
-    }
-    return null;
   }
 
   //-------------- 其他 --------------
